+++ src/components/buttons/genreButton.tsx
@@ -1,7 +1,11 @@
 import { useState } from 'react';
 import { GenreButtonProps } from '@/pages/api/mock';
 
-function GenreButton({ title, selected: initialSelected, editMode }: GenreButtonProps) {
+function GenreButton({
+  title,
+  selected: initialSelected,
+  editMode,
+}: GenreButtonProps) {
   const [selected, setSelected] = useState(initialSelected);
 
   const handleClick = () => {
@@ -13,8 +17,7 @@ function GenreButton({ title, selected: initialSelected, editMode }: GenreButton
       className={`border ${selected ? 'border-black' : 'border-gray-1'} pc:text-14 ${
         editMode ? 'hover:border-green hover:text-green' : ''
       }`}
-      onClick={handleClick}
-    >
+      onClick={handleClick}>
       {title}
     </button>
   );
