+++ src/components/header/index.tsx
@@ -1,13 +1,11 @@
-import Link from 'next/link';
-import CartButton from './cartButton';
-import SearchInput from './searchInput';
-import MyPageButton from './myPageButton';
-import CategoryButton from '@/components/header/categoryButton';
-import SignInButton from './signInButton';
-import SignOutButton from './signOutButton';
-import SignUpButton from './signUpButton';
-import { ReactElement } from 'react';
+import CartButton from '@/components/header/cartButton';
+import SearchInput from '@/components/header/searchInput';
+import MyPageButton from '@/components/header/myPageButton';
+import SignInButton from '@/components/header/signInButton';
+import SignOutButton from '@/components/header/signOutButton';
+import SignUpButton from '@/components/buttons/signUpButton';
 import { HeaderProps } from '@/pages/api/mock';
+import HeaderLayout from '@/containers/HeaderLayout';
 
 function Separator() {
   return (
@@ -19,7 +17,7 @@ function Separator() {
 function NonLoggedInHeader() {
   return (
     <div
-      className="flex tablet:mx-60 justify-between min-w-fit max-w-full items-center h-50 mx-16
+      className="flex tablet:mx-90 justify-between min-w-fit max-w-full items-center h-50 mx-16
         tablet:h-100 pc:h-100 pc:mx-110">
       <div className="text-18 pc:text-24 text-green font-bold tablet:text-24">
         Read Me
@@ -38,7 +36,7 @@ function NonLoggedInHeader() {
 function LoggedInHeader({ numItemsOfCart }: { numItemsOfCart: number }) {
   return (
     <div
-      className="flex tablet:mx-60 justify-between min-w-fit max-w-full items-center h-50 mx-16
+      className="flex tablet:mx-90 justify-between min-w-fit max-w-full items-center h-50 mx-16
         tablet:h-100 pc:h-100 pc:mx-110">
       <div className="text-18 pc:text-24 text-green font-bold tablet:text-24">
         Read Me
@@ -51,35 +49,6 @@ function LoggedInHeader({ numItemsOfCart }: { numItemsOfCart: number }) {
         <CartButton numItemsOfCart={numItemsOfCart} />
         <MyPageButton />
       </div>
-    </div>
-  );
-}
-
-// HeaderTab component
-function HeaderTab() {
-  return (
-    <div className="flex items-center border-gray-3 border-y w-full h-40 tablet:h-70 pc:h-70">
-      <div className="mx-16 tablet:mx-60 pc:mx-110">
-        <CategoryButton />
-      </div>
-      <div
-        className="text-14 tablet:text-16 pc:text-16 inline-flex gap-18 pc:gap-40 tablet:gap-30
-          pc:mx-[3%]">
-        <Link href="/bestsellers"> 베스트</Link>
-        <Link href="/latest"> 신간</Link>
-        <Link href="/custom">맞춤도서</Link>
-        <div className="inline-block border-r w-1 h-14 border-gray-1" />
-        <Link href="/community"> 커뮤니티</Link>
-      </div>
-    </div>
-  );
-}
-
-//헤더 레이아웃
-function HeaderLayout({ children }: { children: ReactElement }) {
-  return (
-    <div className="flex-row h-90 tablet:h-170 pc:h-170">
-      {children} <HeaderTab />
     </div>
   );
 }
